Collect all errors in one module
